Add Optional bind group handles to DrawCall
In order to support arbitrary texture drawing, it is necessary for draw
calls to be able to provide handles to bind groups because that is where
the TextureView and Sampler are bound.

In order to not harm passes which don't need per draw call bind group
handles, the bind group handles that are now optionally provided by the
pass will still be automatically set during each render, but currently
each draw call has the ability to override the bind group handles set,
unsure if this will be an issue or a good thing, time will tell.