Water Complete 2021
I am stamping this commit as it is clear that the engine and the
application code is able to produce good looking low poly water, even if
my numbers are not perfect. Also I think there is some moticum of value
in saying the engine was capable rendering water in 2021, even if it
took several months of procastination to get it ready.

This commit contains a solid ~16 hours of solid effort to get:
* per draw call bind groups (mostly needed for textures)
* basic textured UI rendering
* reflection and refraction passes
* WIP shaders for water

Known issues with the current impl:
* At the moment the water vert surface normal calculation is returning
nan
* The depth calculation for water is not very good, leading to
inconsistent murkiness coloring
* The reflection pass's clip plane is a bit high leading to not very
good reflections for lower height terrain

API issues:
* `push_constant_ranges` ought to be defined in the `PipelineContext`
* `init.rs` is a bit unwieldy, would be nice to split it up further
* There should be some validation on the Vertex `desc` or maybe derive
the label of the Vertex from the Vertex definition.
* May be helpful to use staging belt for Uniform Bind Groups that are
the slightly different between passes/draw calls. Currently the clip
planes are defined as separate buffers with their own Bind Groups