Fix water surface normals & reflections
There were 2 issues manifesting in the previous surface normal
calculation.

* The indicator offset calculation erroniously converted to unsigned
bytes as a joyous side effect of referencing java code in the
implementation. Java `byte`s are implicitly signed :)
* The reference code was using Vec2's and I used Vec3's so the indexing
into the `y` of the vertex instead of the actual `z` value.

Fixed the reflections in the water by properly using math and realizing
that the camera flip assumed the water height = 0. This resulted in the
driver code making some concessions but I am still not sold on the idea
that water *has* to be height = 0, though I don't have a good reason for
it to not be.

Split out the modules back into crates as hopping around the conjoined
file structure wasn't an enjoyable experience for how I write code. Also
took this opportunity to choose another fun name for a crate, `base` /
`core` is now `atlas`. The reasoning for this is a bit of a reach but
given how much of the code is now living in that crate versus the game
client or server crates it seems like a decent description, `atlas` is
holding nearly all the interesting code relating to core functionality.

This commit was supposed to just be water fixes but I forgot to press
the commit button before the crate split but such is life.