Refactor shader uniforms to make more sense
While trying to refactor the water pass to use a model matrix uniform
instead of a push constant to save a bit of processing/data pushing, I
ran into the default limit for `max_bind_groups`, 4, and according to
the documentation[1] it is better to not fiddle with these default
limits as it could limit targeted `device`s.

In order to get around this limitation, I decided to reorganize the bind
groups to be more aligned with this limitation, the  current approach is
to have a bind group for shared world uniforms such as camera and global
lighting, a bind group for texture sampling, and the last two bind
groups can be up to the draw call's discretion.

The API for registering entities with the wrangler changed a bit, now
the `bind_groups` is just a slice of handles rather than an Option of a
slice of handles. The reasoning for this is just personaly preference
but I think down the line more complex structures will more often like
to provide their own bind group. Ideally the backend would reflect this
model a bit more and possibly enforce the max bind groups so I don't
mess up again but that's for another day and it would mostly be hidden
from the user.

To add to this, since push constants are not always a `Mat4` model
matrix, `push_constant_ranges` is now part of `PipelineContext` so each
render pass can define this to fit the use case rather than always
assuming a specific usage.

There are also a few random changes here and there as I go stuck for a
bit trying to work my way through validation errors and some strange
output. I think I will need to create "this thing isn't working, have
you tried these debug steps" runbook at some point. For the 2nd time
this week I ran into bashing my head against sampling a no longer bound
to a color attachment texture, ideally I would remember these need to
have their bind groups recreated on resize, but there needs to be some
sort of fall back to quickly realize the issue rather than getting stuck
forever. Added a note in the `init/texture.rs` file to hopefully save
myself in the future if I run across this.

Reference [1]: https://docs.rs/wgpu/0.12.0/wgpu/struct.Limits.html#structfield.max_bind_groups