Fix water depth calculation in fragment shader
Well this has been a humbling experience through and through. It turns
out when you don't pass the GPU things correctly you get stange
behavior.

The issue I was facing was that the alignment of the camera uniforms
were not correct but also not wrong enough for the `wgpu` validation
layers to save me. So I must now consider how I wish to proceed here, I
would rather explicitly state the `_padding*` to make it less opaque,
rather than implicitly only use 8 and 16 byte aligned types which some
believe to be the more sane approach[1].

Though if this past week of commits has proven anything; it is that I
am anything but that. At the moment I am still thinking over the best
approach but likely I will just do some initial validation on the
alignment but not attempt to do any autocorrection.

I am also going to start a `debug_runbook.md` for when I inevitably end
up in another mind melting scenario that I have put myself in recently.
The 2 immediately obvious matters are going to be texture rebinding on
resize and uniform buffer alignment, especially the last value.

Also this effectively wraps up water as "done", though it isn't perfect
and some values need to be tweaked, it is implemented to "spec"[2][3].

Reference [1]: https://stackoverflow.com/questions/38172696/should-i-ever-use-a-vec3-inside-of-a-uniform-buffer-or-shader-storage-buffer-o
Reference [2]: https://www.youtube.com/watch?v=5yhDb9dzJ58
Reference [3]: https://github.com/TheThinMatrix/LowPolyWater