Switch camera projection to reverse infinite perspective
Following along with "Foundations of Game Engine Development" to
implement global shading passes I wanted to be certain my project matrix
aligns with the one in the book and noticed some interesting tidbits
about depth precision and infinite perspective projection's.

I already don't really have a far clipping plane that is visible so
using an infinite projection makes sense in my case. Something that is
recommended in this chapter is to reverse the depth calculations/buffers
such that the closer depth is 1.0 and the furthest is 0.0, this is due
to how floating point precision works, short of it is that nearly all of
the precision is between 0.0 and 0.5, with only a small fraction of the
precision being avaliable to 0.5 through 1.0. Lots of great free
resources on this [1][2][3].

Anyways, I have used up my alotted math processing power the day and
this project is working with reversed infinite perspective projection on
the camera.

Reference [1]: https://developer.nvidia.com/content/depth-precision-visualized
Reference [2]: http://www.geometry.caltech.edu/pubs/UD12.pdf
Reference [3]: https://outerra.blogspot.com/2012/11/maximizing-depth-buffer-range-and.html