Start implementing init code for global shadows
Nearly there, just need a bit of a refactor (read: API churn) on the
Pass's to be able to support indexed texture views for our 2D depth
texture.

Two thoughts, either refactor the `Texture` struct to support multiple
views, or completely remove the view book keeping off of `Texture` but
that adds the headache of what happens at resize.