CPU side optimization pass
Allocate wrangler resources on stack using `smallvec`.

Attempt to add SIMD to math lib, current state is that `Vec2` and `Vec4`
see either no improvement over non-explicit SIMD (not sure how
LLVM/rustc optimizes) in most cases. `Vec3` sees similar improvement as
`Vec4` with drawback of ruining dispatching vertex data to GPU by adding
a bunch of junk bytes, also segfaults when I move the camera so I am
taking out the SIMD impl for `Vec3` for now.

The only reason I am leaving SIMD code in is because my Vec mulitply
benchmarks are currently seeing a speedup of ~2x in SIMD vs non SIMD,
but I am unable to reproduce this speedup in more complex cases so this
may just be smoke in mirrors or something. Example of speedup
```
vec4 multiply vecs      time:   [212.75 ps 212.79 ps 212.84 ps]
                        change: [-49.767% -49.698% -49.638%] (p = 0.00 < 0.05)
                        Performance has improved.
```
Granted this is in picoseconds so we are acutely at the mercy of
electrons here.